mob next [ci-skip] [ci skip] [skip ci]
lastFile:webapi/src/main/scala/org/knora/webapi/routing/admin/ProjectsRouteZ.scala
