+++ webapi/src/main/scala/org/knora/webapi/core/HttpServerWithZIOHttp.scala
@@ -4,100 +4,23 @@
  */
 
 package org.knora.webapi.core
-import akka.actor.ActorRef
-import akka.pattern.ask
-import akka.util.Timeout
-import org.knora.webapi.config.AppConfig
-import org.knora.webapi.messages.admin.responder.projectsmessages.{
-  ProjectGetRequestADM,
-  ProjectGetResponseADM,
-  ProjectIdentifierADM
-}
-import org.knora.webapi.messages.util.KnoraSystemInstances
-import org.knora.webapi.routing.HealthRouteWithZIOHttp
-import zhttp.http._
 import zhttp.service.Server
-import zio.json.DeriveJsonEncoder
-import zio.{ZLayer, _}
+import zio.ZLayer
+import zio._
 
-import java.net.URLDecoder
-// import java.net.URLDecoder
-// import dsp.config.AppConfig
-
-case class HelloZio(hello: String)
-
-object HelloZio {
-  implicit val encoder = DeriveJsonEncoder.gen[HelloZio]
-}
-
-final case class HelloZioApp(router: AppRouter, appConfig: AppConfig) {
-  implicit val sender: ActorRef = router.ref
-  implicit val timeout: Timeout = appConfig.defaultTimeoutAsDuration
-
-  /**
-   * Returns a single project identified through the IRI.
-   * //
-   */
-//  private def getProjectByIri(): Route =
-//    path(projectsBasePath / "iri" / Segment) { value =>
-//      get { requestContext =>
-//        val requestMessage: Future[ProjectGetRequestADM] = for {
-//          requestingUser <- getUserADM(
-//            requestContext = requestContext,
-//            routeData.appConfig
-//          )
-//
-//        } yield ProjectGetRequestADM(
-//          identifier = IriIdentifier
-//            .fromString(value)
-//            .getOrElseWith(e => throw BadRequestException(e.head.getMessage)),
-//          requestingUser = requestingUser
-//        )
-//
-//        RouteUtilADM.runJsonRoute(
-//          requestMessageF = requestMessage,
-//          requestContext = requestContext,
-//          appActor = appActor,
-//          log = log
-//        )
-//      }
-//    }
-  def route(): HttpApp[Any, Nothing] =
-    Http.collectZIO[Request] {
-
-      // GET admin/projects/iri/{iri}
-      case Method.GET -> !! / "admin" / "projects" / "iri" / iri =>
-        for {
-          user       <- ZIO.succeed(KnoraSystemInstances.Users.SystemUser)
-          iriDecoded <- ZIO.attempt(URLDecoder.decode(iri, "utf-8")).orDie
-          iriValue <- ProjectIdentifierADM.IriIdentifier
-                        .fromString(iriDecoded)
-                        .toZIO
-                        .orDie
-          message = ProjectGetRequestADM(
-                      identifier = iriValue,
-                      requestingUser = user
-                    )
-          response <- ZIO.fromFuture(_ => router.ref.ask(message)).map(_.asInstanceOf[ProjectGetResponseADM]).orDie
-        } yield Response.json(response.toJsValue.toString())
-
-    }
-}
-object HelloZioApp {
-  val layer: ZLayer[AppRouter with AppConfig, Nothing, HelloZioApp] = ZLayer.fromFunction { (router, config) =>
-    HelloZioApp(router, config)
-  }
-}
+import org.knora.webapi.config.AppConfig
+import org.knora.webapi.routing.HealthRouteWithZIOHttp
+import org.knora.webapi.routing.admin.ProjectsRouteZ
 
 object HttpServerWithZIOHttp {
 
   val routes = HealthRouteWithZIOHttp()
 
-  val layer: ZLayer[AppRouter & AppConfig & State & HelloZioApp, Nothing, Unit] =
+  val layer: ZLayer[AppRouter & AppConfig & State & ProjectsRouteZ, Nothing, Unit] =
     ZLayer {
       for {
         appConfig   <- ZIO.service[AppConfig]
-        helloZioApp <- ZIO.service[HelloZioApp]
+        helloZioApp <- ZIO.service[ProjectsRouteZ]
         r            = routes ++ helloZioApp.route()
         port         = appConfig.knoraApi.externalZioPort
         _           <- Server.start(port, r).forkDaemon
